Convert BuildBucketProvider into a build service
