+++ build-logic/integration-testing/src/main/kotlin/gradlebuild.cross-version-tests.gradle.kts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import gradlebuild.integrationtests.TestType
+import gradlebuild.testing.TestType
 import gradlebuild.integrationtests.addDependenciesAndConfigurations
 import gradlebuild.integrationtests.addSourceSet
 import gradlebuild.integrationtests.configureIde
